refactor(snapping): Move to selector
