+++ lib/components/modification/edit-alignment.js
@@ -14,7 +14,7 @@ import colors from '../../constants/colors'
 import {Button} from '../buttons'
 import {Checkbox, Number} from '../input'
 import type {MapState, GTFSStop} from '../../types'
-import getExistingStopsAlongPattern from '../../utils/get-existing-stops-along-pattern'
+import getExistingStopsAlongPattern from '../../selectors/existing-stops-along-pattern'
 
 type Props = {
   disabled: boolean,
@@ -216,7 +216,7 @@ export default class EditAlignment extends PureComponent {
           <Button block
             onClick={this._autoGen}
             style='warning'>
-            <Icon type='circle' /> message('transitEditor.snap')
+            <Icon type='circle' /> {message('transitEditor.snap')}
           </Button>
         }
