+++ lib/selectors/existing-stops-along-pattern.js
